feat(runners)!: process multiple events at once (#4)
* feat(control-plane): add support for handling multiple events in a single invocation

Currently we restrict the `scale-up` Lambda to only handle a single
event at a time. In very busy environments this can prove to be a
bottleneck: there are calls to GitHub and AWS APIs that happen each
time, and they can end up taking long enough that we can't process
job queued events faster than they arrive.

In our environment we are also using a pool, and typically we have
responded to the alerts generated by this (SQS queue length growing) by
expanding the size of the pool. This helps because we will more
frequently find that we don't need to scale up, which allows the lambdas
to exit a bit earlier, so we can get through the queue faster. But it
makes the environment much less responsive to changes in usage patterns.

At its core, this Lambda's task is to construct an EC2 `CreateFleet`
call to create instances, after working out how many are needed. This is
a job that can be batched. We can take any number of events, calculate
the diff between our current state and the number of jobs we have,
capping at the maximum, and then issue a single call.

The thing to be careful about is how to handle partial failures, if EC2
creates some of the instances we wanted but not all of them. Lambda has
a configurable function response type which can be set to
`ReportBatchItemFailures`. In this mode, we return a list of failed
messages from our handler and those are retried. We can make use of this
to give back as many events as we failed to process.

Now we're potentially processing multiple events in a single Lambda, one
thing we should optimise for is not recreating GitHub API clients. We
need one client for the app itself, which we use to find out
installation IDs, and then one client for each installation which is
relevant to the batch of events we are processing. This is done by
creating a new client the first time we see an event for a given
installation.

We also remove the same `batch_size = 1` constraint from the `job-retry`
Lambda and make it configurable instead, using AWS's default of 10 for
SQS if not configured. This Lambda is used to retry events that
previously failed. However, instead of reporting failures to be retried,
here we maintain the pre-existing fault-tolerant behaviour where errors
are logged but explicitly do not cause message retries, avoiding
infinite loops from persistent GitHub API issues or malformed events.

Tests are added for all of this.

* handle ScaleError

* fix: correct typo in test comment

* style: minor tweaks to `ScaleError`

Some small tweaks to the style of `ScaleError`, while we're working on
it.

`message` doesn't need to be a public property. `Error` already has it.
Likewise with `stack` - we can use the parent's implementation.

Instead of `getDetailedMessage` we can have a property getter which is
very slightly nicer to read at callsites (`e.detailedMessage` vs.
`e.getDetailedMessage()`).

* style: use guard clauses in `processFleetResult`

This function is getting a little complex to read now. Refactoring to
use the guard clause pattern makes the flow clearer. Now it's more like
a sequence of checks followed by a final `throw`.

* refactor: let `ScaleError` convert itself to batch item failures, add tests

This lets us move some logic from `processFleetResult`, simplifying it.
We're additionally adding some bounds checking here, and testing it all.

* refactor: hardcode ScaleError message

We only construct this once in the production code - there's no need to
have it be variable.

* chore: fix tpo nur -> number

* fix: only retry when we don't create enough ephemeral runners

Prior to this change, we would have retried if _any_ runner failed to be
created either due to capacity or an EC2 error. Now we'll consider these
differently.

If we can't create due to the maximum allowed runner count, retry only
if creating ephemeral runners.

Always retry if EC2 creation fails.

* style: combine two nested `if`s

The code:

```typescript
if (x) {
  if (y) {
    // ...
  }
}
```

is equivalent to

```typescript
if (x && y) {
  // ...
}
```

the latter is shorter and easier to read, so let's use that.

* fix: avoid grouping all events

We were grouping all events we were called for, and only looking at the
non-`aws:sqs` events to warn about them. This can be done as we go, and
then we only need to keep the SQS events in memory.

* chore(scale-runners): log failures with same key name as variable name

Copilot pointed out correctly that it could be confusing to log
`failedInstanceCount` using the key `missingInstanceCount` when there is
a variable with that name.

---------

Co-authored-by: Iain Lane <iain@grafana.com>